+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Object Lifecycle Management support for Multipart Uploads Cleanup
 - Support for Autonomous JSON database added
+- Support resource discovery for Blockchain resources
 
 ## 3.88.0 (August 05, 2020)